fix: Address PR #21 review comments
全ての未解決レビューコメントに対応:

1. docs/WORKFLOW_IMPROVEMENT_PLAN.md をphase3ディレクトリへ移動
   - Phase依存の計画なので適切な場所に配置

2. packages/core/src/workflows/impl-functional/index.ts のコメント修正
   - default-workflows.jsのエクスポートは必要なので維持
   - コメントを正確な内容に修正

3. packages/shared-types/src/workflow.ts の型を改善
   - any型を削除し、適切な型を指定
   - createDriverの戻り値をAIDriverに変更
   - 不要なLLMOptions定義を削除
   - @moduler-prompt/driverのAIDriverを直接使用

4. prompts/BUILD_WORKFLOW.md を削除
   - 改善済みの内容なので削除

5. extended-workflows.ts を default-workflows.ts にリネーム
   - より適切な名前に変更
   - インポートパスも更新
