+++ packages/core/src/workflows/impl-functional/index.ts
@@ -1,8 +1,37 @@
-// 全ワークフロー定義をエクスポート
+/**
+ * デフォルトワークフローコレクション
+ */
+
+import type { WorkflowDefinition, IWorkflowRegistry } from '../workflow-types.js';
+import { LogType } from '../logger.js';
+import { ingestInputWorkflow } from './ingest-input.js';
+import { processUserRequestWorkflow } from './process-user-request.js';
+import { analyzeIssueImpactWorkflow } from './analyze-issue-impact.js';
+import { extractKnowledgeWorkflow } from './extract-knowledge.js';
+
+// 全ワークフロー定義を再エクスポート
 export { ingestInputWorkflow } from './ingest-input.js';
 export { processUserRequestWorkflow } from './process-user-request.js';
 export { analyzeIssueImpactWorkflow } from './analyze-issue-impact.js';
 export { extractKnowledgeWorkflow } from './extract-knowledge.js';
 
-// デフォルトワークフローと登録ヘルパー
-export * from './extended-workflows.js';
+/**
+ * 全てのデフォルトワークフロー
+ */
+export const DEFAULT_WORKFLOWS: WorkflowDefinition[] = [
+  ingestInputWorkflow,
+  processUserRequestWorkflow,
+  analyzeIssueImpactWorkflow,
+  extractKnowledgeWorkflow,
+];
+
+/**
+ * デフォルトワークフローを登録するヘルパー関数
+ */
+export function registerDefaultWorkflows(registry: IWorkflowRegistry): void {
+  for (const workflow of DEFAULT_WORKFLOWS) {
+    registry.register(workflow);
+  }
+  console.log(`Registered ${DEFAULT_WORKFLOWS.length} default workflows`);
+}
+
